+++ server/api/community_wraps.go
@@ -174,6 +174,11 @@ func (h *handlers) communityWrapsUpload(w http.ResponseWriter, r *http.Request) 
 	req.Header.Set("Content-Type", writer.FormDataContentType())
 	req.Header.Set("X-Fingerprint", getFingerprint())
 
+	// Forward passcode if present (admin bypasses rate limiting)
+	if passcode := r.Header.Get("X-Passcode"); passcode != "" {
+		req.Header.Set("X-Passcode", passcode)
+	}
+
 	client := &http.Client{Timeout: 30 * time.Second}
 	resp, err := client.Do(req)
 	if err != nil {
@@ -208,6 +213,11 @@ func (h *handlers) communityWrapsDownload(w http.ResponseWriter, r *http.Request
 		return
 	}
 	req.Header.Set("X-Fingerprint", getFingerprint())
+
+	// Forward passcode if present (admin bypasses rate limiting)
+	if passcode := r.Header.Get("X-Passcode"); passcode != "" {
+		req.Header.Set("X-Passcode", passcode)
+	}
 
 	client := &http.Client{Timeout: 30 * time.Second}
 	resp, err := client.Do(req)
