+++ web/src/pages/CommunityWraps.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState, useCallback, useRef } from "react"
-import { Search, Upload, Download, Paintbrush, ChevronLeft, ChevronRight, Loader2, CheckCircle, AlertCircle, Trash2, Pencil, Shield, X } from "lucide-react"
+import { Search, Upload, Download, Paintbrush, ChevronLeft, ChevronRight, Loader2, CheckCircle, AlertCircle, Trash2, Pencil } from "lucide-react"
 import GodotRenderer, { type GodotRendererHandle } from "../components/wraps/GodotRenderer"
 
 const API_BASE = "/api"
@@ -64,42 +64,12 @@ interface LibraryResponse {
 
 type Tab = "browse" | "upload"
 
-export default function CommunityWraps({ onRegisterHeadingClick }: { onRegisterHeadingClick?: (fn: () => void) => void } = {}) {
+export default function CommunityWraps({ adminPasscode, onAdminPasscodeChange }: { adminPasscode: string | null; onAdminPasscodeChange: (v: string | null) => void }) {
   const [tab, setTab] = useState<Tab>("browse")
-  const [adminPasscode, setAdminPasscode] = useState<string | null>(null)
-  const [showPasscodePrompt, setShowPasscodePrompt] = useState(false)
-  const clickCountRef = useRef(0)
-  const clickTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null)
 
   // Godot 3D engine state — mounted at page level so it starts loading immediately
   const godotReadyRef = useRef(false)
   const godotRef = useRef<GodotRendererHandle>(null)
-
-  const handleHeadingClick = () => {
-    if (adminPasscode) {
-      // Already in admin mode — 5 clicks exits
-      clickCountRef.current++
-      if (clickTimerRef.current) clearTimeout(clickTimerRef.current)
-      clickTimerRef.current = setTimeout(() => { clickCountRef.current = 0 }, 2000)
-      if (clickCountRef.current >= 5) {
-        clickCountRef.current = 0
-        setAdminPasscode(null)
-      }
-      return
-    }
-
-    clickCountRef.current++
-    if (clickTimerRef.current) clearTimeout(clickTimerRef.current)
-    clickTimerRef.current = setTimeout(() => { clickCountRef.current = 0 }, 2000)
-    if (clickCountRef.current >= 5) {
-      clickCountRef.current = 0
-      setShowPasscodePrompt(true)
-    }
-  }
-
-  useEffect(() => {
-    onRegisterHeadingClick?.(handleHeadingClick)
-  }) // re-register on every render so closure stays fresh
 
   return (
     <div className="space-y-6">
@@ -126,15 +96,6 @@ export default function CommunityWraps({ onRegisterHeadingClick }: { onRegisterH
           Upload
         </button>
 
-        {adminPasscode && (
-          <div className="ml-auto flex items-center gap-1.5 rounded bg-red-500/10 border border-red-500/20 px-2.5 py-1 text-xs text-red-400">
-            <Shield className="h-3 w-3" />
-            Admin Mode
-            <button onClick={() => setAdminPasscode(null)} className="ml-1 hover:text-red-300">
-              <X className="h-3 w-3" />
-            </button>
-          </div>
-        )}
       </div>
 
       {/* Hidden Godot renderer — starts loading 283MB .pck immediately */}
@@ -147,89 +108,11 @@ export default function CommunityWraps({ onRegisterHeadingClick }: { onRegisterH
       />
 
       {tab === "browse" ? (
-        <BrowseTab adminPasscode={adminPasscode} onAdminExit={() => setAdminPasscode(null)} />
+        <BrowseTab adminPasscode={adminPasscode} onAdminExit={() => onAdminPasscodeChange(null)} />
       ) : (
         <UploadTab godotReadyRef={godotReadyRef} godotRef={godotRef} adminPasscode={adminPasscode} />
       )}
 
-      {/* Passcode prompt modal */}
-      {showPasscodePrompt && (
-        <PasscodeModal
-          onSuccess={(passcode) => {
-            setAdminPasscode(passcode)
-            setShowPasscodePrompt(false)
-          }}
-          onClose={() => setShowPasscodePrompt(false)}
-        />
-      )}
-    </div>
-  )
-}
-
-function PasscodeModal({ onSuccess, onClose }: { onSuccess: (passcode: string) => void; onClose: () => void }) {
-  const [input, setInput] = useState("")
-  const [error, setError] = useState<string | null>(null)
-  const [validating, setValidating] = useState(false)
-
-  const handleValidate = async () => {
-    if (!input.trim()) return
-    setValidating(true)
-    setError(null)
-    try {
-      const res = await fetch(`${API_BASE}/wraps/admin/validate`, {
-        method: "POST",
-        headers: { "x-passcode": input.trim() },
-      })
-      if (res.ok) {
-        onSuccess(input.trim())
-      } else {
-        setError("Invalid passcode")
-        setInput("")
-      }
-    } catch {
-      setError("Connection failed")
-    } finally {
-      setValidating(false)
-    }
-  }
-
-  return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/60 p-4" onClick={onClose}>
-      <div
-        className="w-full max-w-sm overflow-hidden rounded-2xl border border-white/10 bg-slate-900 p-6"
-        onClick={(e) => e.stopPropagation()}
-      >
-        <h3 className="text-lg font-semibold text-slate-100">Admin Access</h3>
-        <p className="mt-1 text-xs text-slate-500">Enter the admin passcode to continue</p>
-        <input
-          type="password"
-          value={input}
-          onChange={(e) => setInput(e.target.value)}
-          onKeyDown={(e) => e.key === "Enter" && handleValidate()}
-          placeholder="Passcode"
-          autoFocus
-          className="mt-4 w-full rounded-lg border border-white/10 bg-white/[0.03] px-3 py-2 text-sm text-slate-200 placeholder:text-slate-600 focus:border-blue-500/50 focus:outline-none"
-        />
-        {error && (
-          <p className="mt-2 text-xs text-red-400">{error}</p>
-        )}
-        <div className="mt-4 flex gap-3">
-          <button
-            onClick={handleValidate}
-            disabled={!input.trim() || validating}
-            className="flex flex-1 items-center justify-center gap-2 rounded-lg bg-blue-600 px-4 py-2.5 text-sm font-medium text-white transition-colors hover:bg-blue-500 disabled:opacity-50"
-          >
-            {validating && <Loader2 className="h-4 w-4 animate-spin" />}
-            Validate
-          </button>
-          <button
-            onClick={onClose}
-            className="rounded-lg border border-white/10 px-4 py-2.5 text-sm text-slate-400 transition-colors hover:bg-white/5"
-          >
-            Cancel
-          </button>
-        </div>
-      </div>
     </div>
   )
 }
@@ -286,7 +169,9 @@ function BrowseTab({ adminPasscode, onAdminExit }: { adminPasscode: string | nul
   const handleDownload = async (wrap: CommunityWrap) => {
     setDownloading(wrap.code)
     try {
-      const res = await fetch(`${API_BASE}/wraps/download/${wrap.code}`, { method: "POST" })
+      const headers: Record<string, string> = {}
+      if (adminPasscode) headers["x-passcode"] = adminPasscode
+      const res = await fetch(`${API_BASE}/wraps/download/${wrap.code}`, { method: "POST", headers })
       if (!res.ok) {
         const data = await res.json().catch(() => ({}))
         throw new Error(data.error || `HTTP ${res.status}`)
