+++ web/src/pages/LockChime.tsx
@@ -96,10 +96,8 @@ async function getWavDuration(file: File): Promise<number> {
 // Main component
 // ─────────────────────────────────────────────────────────────
 
-export default function LockChime() {
+export default function LockChime({ adminPasscode, onAdminPasscodeChange }: { adminPasscode: string | null; onAdminPasscodeChange: (v: string | null) => void }) {
   const [tab, setTab] = useState<Tab>("library")
-  const [adminPasscode, setAdminPasscode] = useState<string | null>(null)
-  const [showPasscodePrompt, setShowPasscodePrompt] = useState(false)
   const [volume, setVolume] = useState(() => {
     const saved = localStorage.getItem("lockchime-preview-volume")
     return saved !== null ? Number(saved) : 0.5
@@ -134,15 +132,6 @@ export default function LockChime() {
           Community
         </button>
 
-        {adminPasscode && (
-          <div className="ml-auto flex items-center gap-1.5 rounded bg-red-500/10 border border-red-500/20 px-2.5 py-1 text-xs text-red-400">
-            <Shield className="h-3 w-3" />
-            Admin Mode
-            <button onClick={() => setAdminPasscode(null)} className="ml-1 hover:text-red-300">
-              <X className="h-3 w-3" />
-            </button>
-          </div>
-        )}
       </div>
 
       {/* USB Disconnect Notice */}
@@ -178,89 +167,10 @@ export default function LockChime() {
         <CommunityTab adminPasscode={adminPasscode} volume={volume} />
       )}
 
-      {/* Passcode modal */}
-      {showPasscodePrompt && (
-        <PasscodeModal
-          onSuccess={(passcode) => {
-            setAdminPasscode(passcode)
-            setShowPasscodePrompt(false)
-          }}
-          onClose={() => setShowPasscodePrompt(false)}
-        />
-      )}
     </div>
   )
 }
 
-// ─────────────────────────────────────────────────────────────
-// Passcode modal (reuses the same pattern as Community Wraps)
-// ─────────────────────────────────────────────────────────────
-
-function PasscodeModal({ onSuccess, onClose }: { onSuccess: (passcode: string) => void; onClose: () => void }) {
-  const [input, setInput] = useState("")
-  const [error, setError] = useState<string | null>(null)
-  const [validating, setValidating] = useState(false)
-
-  const handleValidate = async () => {
-    if (!input.trim()) return
-    setValidating(true)
-    setError(null)
-    try {
-      const res = await fetch(`${API_BASE}/lockchime/community/admin/validate`, {
-        method: "POST",
-        headers: { "x-passcode": input.trim() },
-      })
-      if (res.ok) {
-        onSuccess(input.trim())
-      } else {
-        setError("Invalid passcode")
-        setInput("")
-      }
-    } catch {
-      setError("Connection failed")
-    } finally {
-      setValidating(false)
-    }
-  }
-
-  return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/60 p-4" onClick={onClose}>
-      <div
-        className="w-full max-w-sm overflow-hidden rounded-2xl border border-white/10 bg-slate-900 p-6"
-        onClick={(e) => e.stopPropagation()}
-      >
-        <h3 className="text-lg font-semibold text-slate-100">Admin Access</h3>
-        <p className="mt-1 text-xs text-slate-500">Enter the admin passcode to continue</p>
-        <input
-          type="password"
-          value={input}
-          onChange={(e) => setInput(e.target.value)}
-          onKeyDown={(e) => e.key === "Enter" && handleValidate()}
-          placeholder="Passcode"
-          autoFocus
-          className="mt-4 w-full rounded-lg border border-white/10 bg-white/[0.03] px-3 py-2 text-sm text-slate-200 placeholder:text-slate-600 focus:border-violet-500/50 focus:outline-none"
-        />
-        {error && <p className="mt-2 text-xs text-red-400">{error}</p>}
-        <div className="mt-4 flex gap-3">
-          <button
-            onClick={handleValidate}
-            disabled={!input.trim() || validating}
-            className="flex flex-1 items-center justify-center gap-2 rounded-lg bg-violet-600 px-4 py-2.5 text-sm font-medium text-white transition-colors hover:bg-violet-500 disabled:opacity-50"
-          >
-            {validating && <Loader2 className="h-4 w-4 animate-spin" />}
-            Validate
-          </button>
-          <button
-            onClick={onClose}
-            className="rounded-lg border border-white/10 px-4 py-2.5 text-sm text-slate-400 transition-colors hover:bg-white/5"
-          >
-            Cancel
-          </button>
-        </div>
-      </div>
-    </div>
-  )
-}
 
 // ─────────────────────────────────────────────────────────────
 // Toast component (shared)
@@ -1352,7 +1262,9 @@ function CommunityBrowse({ adminPasscode, volume }: { adminPasscode: string | nu
   async function handleDownload(sound: CommunitySound) {
     setDownloadingCode(sound.code)
     try {
-      const res = await fetch(`${API_BASE}/lockchime/community/download/${sound.code}`, { method: "POST" })
+      const headers: Record<string, string> = {}
+      if (adminPasscode) headers["x-passcode"] = adminPasscode
+      const res = await fetch(`${API_BASE}/lockchime/community/download/${sound.code}`, { method: "POST", headers })
       const data = await res.json().catch(() => ({}))
       if (!res.ok) throw new Error(data.error || `HTTP ${res.status}`)
       showToast(`Downloaded "${sound.name}" to your library`, "success")
